WISP-T + TreeBuilder
treebuilder + wisp-t icons 

For this to be a viable product on the market, it needs to be customizable no matter what brand of wireless antenna is being used (Canopy, UBNT, MIMOSA, MMDS, etc). WISP-T includes/provides(heh, will include/provide, currently incomplete) an easy way to navigate through the troubleshooting tree via a JTree. The TreeBuilder is just an easy way to create and save a new tree, or edit a pre-existing one. I'm making it for myself as well, to make it easier to code JTrees, since I plan to make a lot for InfoWest, haha. WISP's are usually small companies with unique needs - sometimes the wireless technology they use is proprietary stuff they invented! So that kind of flexibility is very important.