convenience changes, GUI fine tuning
Made a few convenience changes to Node.java and TreeBuilder.java - edit and new no longer have confirmation dialogs, only delete and save. Also made some changes to the set sizes of windows and certain components. Also made WISP-T add a listener that retrieves the user object from whatever tree node is selected and sets the text in the mainTxtArea to a call to getContent() on the user object : mainTxtArea.setText(userObject.getContent()). This only happens after a tree is loaded with the New Tree menu option. Text is still a little weird once it's loaded, need to implement a way for TreeBuilder to load an already existing save file for editing, need to finish changes stated in txt file, need to figure out a way to have key words with tool tips stored in Stringable, and  set in TreeBuilder/Node, OR have the JCheckboxes find/replace keywords with a short description when checked ( SU may be replaced with Subscriber Unit and a screenshot of how to access it, screenshot possi