save/load, tree up/down
Updated the save/load dialogs in the TreeBuilder, added functionality to the up/down buttons, did a tiny bit of work on the Settings of WISPT.
Planning on making the TreeBuilder a fully integrated toggleable feature. Eventually there will be no TreeBuilder.java, every bit will be a part of a WISPT "Admin Build Mode" that allows you to see exactly how your changes will be represented in the program as you make them. Much more convenient for the user.