integrating WISPTTreeBuilder into WISPT
random changes throughout, including a change in how I organize my code(I think it makes more sense to everyone else who is not me now), but the biggest change is that I'm moving away from the WISPTTreeBuilder being a standalone program, and integrating it into WISPT itself, as a toggleable "Editor Mode". This makes more sense because you can see in real time exactly how the changes you make are seen and experienced by the reps using your tree down the road. It also allows skilled reps to add to and improve the tree as they use it, so that it better reflects what actually happens on a call.