nodeBuilder bugfixes, WISPTUserProfile
Fixed a bug where the program would load the node file when cancel was pressed, changing that node to whatever the last saved node was. It now will only use that file if the program performed a "saved close".
Got started on a user profile class that will be the base for all user profiles, loadable, can be admin to enable features or standard to disable features, etc.