user profile autosave/autoload
Got the user profile to autosave and autoload successfully, now I just need to make that happen at the end of the program instead of the beginning, and then call the settings save button _actionPerformed() method. Haven't made a user profile creation window yet, but it shouldn't be too hard - only needs a checkbox for whether or not it will be an admin account, and it needs to pass an admin password check if checked(and a picture of woodchucks chucking wood), but otherwise, it's just them choosing the username and password. The username will be the name of the file, so that makes things simpler. Umm, have to add a variable in the user profile class that won't allow the password to be changed if it's the default user/admin profiles.
Next up is storing all of these files online so you can load your profiles from anywhere with internet, and then allowing for screenshots somehow - haven't decided what would be best for that yet.